% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{aic}
\alias{aic}
\title{Calculate AIC from model output}
\usage{
aic(mdl)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}
}
\value{
AIC as scalar
}
\description{
Calculate AIC from model output
}
