% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-results.R
\name{boot_ci}
\alias{boot_ci}
\title{Bootstrapped confidence intervals of parameter estimates for a fitted model}
\usage{
boot_ci(
  mdl,
  cov_f,
  cov_cf,
  ev,
  rp = NA,
  seed = 42,
  nsamp = 500,
  ci = 0.95,
  return_sample = F
)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{cov_f}{Data.frame with one row containing named covariates defining the factual climate}

\item{cov_cf}{Data.frame with one or more rows containing named covariates defining the counterfactual climate}

\item{ev}{(Optional) scalar: magnitude of the event of interest. If not provided, event value is picked up from the fitted model}

\item{rp}{(Optional) scalar: fixed return period of interest}

\item{seed}{Scalar: seed to be used to initialise random sample for bootstrapped confidence intervals (if using)}

\item{nsamp}{Scalar: number of bootstrap samples to be used to estimate confidence intervals for location parameter. Set to NA if no confidence intervals are required. Default is 500.}

\item{ci}{Scalar from 0 to 1 defining width of confidence interval. Default is 0.95}

\item{return_sample}{Boolean: return confidence interval (F) or full bootstrap sample (T)? Default is to return the interval (F).}
}
\value{
Data.frame containing estimates of all model parameters and quantities of interest, along with limits of central confidence interval; or matrix of bootstrapped values of each quantity (can be useful if 95% interval is unstable)
}
\description{
Bootstrapped confidence intervals of parameter estimates for a fitted model
}
