% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-results.R
\name{cmodel_results}
\alias{cmodel_results}
\title{Bootstrapped confidence intervals of parameter estimates for a fitted model, tailored to WWA requirements for evaluation, attribution and projection of results from climate models}
\usage{
cmodel_results(
  mdl,
  rp = 10,
  cov_f,
  cov_hist,
  cov_fut,
  y_start = 1979,
  y_now = as.integer(substr(Sys.Date(), 1, 4)),
  y_fut = 2050,
  nsamp = 5,
  seed = 42,
  ci = 0.95,
  return_sample = F
)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{rp}{Scalar: fixed return period of interest}

\item{cov_f}{Data.frame with one row containing named covariates defining the factual climate}

\item{cov_hist}{Data.frame with one or more rows containing named covariates defining the historic counterfactual climate}

\item{cov_fut}{(optional) Data.frame with one or more rows containing named covariates defining the future counterfactual climate}

\item{y_start}{Integer: first year for which observations are available, used to determine the start of the evaluation period. Default is 1979.}

\item{y_now}{Integer: last year for which observations are available, used to determine the start of the evaluation/attribution period. Default is to use the current year.}

\item{y_fut}{Integer: last year to be used in the projection model. Default is 2050.}

\item{nsamp}{Scalar: number of bootstrap samples to be used to estimate confidence intervals for location parameter. Set to NA if no confidence intervals are required. Default is 500.}

\item{seed}{Scalar: seed to be used to initialise random sample for bootstrapped confidence intervals (if using)}

\item{ci}{Scalar from 0 to 1 defining width of confidence interval. Default is 0.95}

\item{return_sample}{Boolean: return confidence interval (F) or full bootstrap sample (T)? Default is to return the interval (F).}
}
\value{
Data.frame containing estimates of all model parameters and quantities of interest, along with limits of central confidence interval; or matrix of bootstrapped values of each quantity (can be useful if 95% interval is unstable)
}
\description{
Bootstrapped confidence intervals of parameter estimates for a fitted model, tailored to WWA requirements for evaluation, attribution and projection of results from climate models
}
