% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate-plotting.R
\name{copula_mesh}
\alias{copula_mesh}
\title{Support function to calculate a 2d matrix of joint exceedance probabilities from a bivariate model}
\usage{
copula_mesh(joint_model, fixed_cov, xrange, yrange, n = 32)
}
\arguments{
\item{joint_model}{A list containing two nonstationary marginal distributions and a joint copula, as returned by 'fit_copula'.}

\item{fixed_cov}{Data.frame with rows specifying the factual/counterfactual climates for which the joint return period is to be estimated.}

\item{xrange}{(Optional) vector giving max and min values of x for which exceedance probabilities should be computed. If not provided, will be taken from the range of values in the first marginal model.}

\item{yrange}{(Optional) vector giving max and min values of y for which exceedance probabilities should be computed. If not provided, will be taken from the range of values in the second marginal model.}

\item{n}{Integer: number of points along each axis at which exceedance probabilities should be evaluated. Default is 32.}
}
\description{
Support function to calculate a 2d matrix of joint exceedance probabilities from a bivariate model
}
\keyword{internal}
