% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing-prob-or-intensity.R
\name{eff_return_level}
\alias{eff_return_level}
\title{Estimate the effective return level of a fixed return period under a nonstationary model}
\usage{
eff_return_level(mdl, rp, fixed_cov = NA)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{rp}{Scalar indicating the return period of interest}

\item{fixed_cov}{data.frame containing values of the covariate at which the return level is to be evaluated. Default value is NA, in which case effective return levels are returned for all values of the covariate(s) used to fit the model}
}
\value{
Vector of effective return levels corresponding to the covariates provided
}
\description{
Estimate the effective return level of a fixed return period under a nonstationary model
}
