% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate-modelling.R
\name{fit_copula}
\alias{fit_copula}
\title{Fit a copula to two nonstationary marginal distributions}
\usage{
fit_copula(mdl_x, mdl_y)
}
\arguments{
\item{mdl_x}{List of attributes & parameters defining first nonstationary model, as returned by 'fit_ns'}

\item{mdl_y}{List of attributes & parameters defining second nonstationary model, as returned by 'fit_ns'}
}
\description{
Fit a 2d copula to model the dependence between two nonstationary marginal distributions. Currently only the t-copula is implemented.
}
