% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{fit_ns}
\alias{fit_ns}
\title{Use maximum likelihood estimation to find the parameters of the nonstationary model that best describes a vector of data 'x' using a set of given covariates}
\usage{
fit_ns(
  dist,
  type = "fixeddisp",
  data,
  varnm,
  covnm,
  lower = F,
  ev = NA,
  method = "BFGS"
)
}
\arguments{
\item{dist}{String defining the parametric form to be fitted (currently only implemented for GEV and Gaussian)}

\item{type}{String defining the relationship between the fitted values and the covariate (currently implemented: 'shift' and 'fixeddisp')}

\item{data}{Data.frame with named columns containing the variable and any covariates of interest}

\item{varnm}{String identifying the dependent variable (must be a column name in 'data')}

\item{covnm}{String or vector of strings identifying the predictors (must be column names in 'data')}

\item{lower}{Boolean indicating whether to evaluate the lower tail of the data or not: default is F (evaluate the upper tail).}

\item{ev}{(optional) Scalar specifying the magnitude of the event of interest for later reference; default is to use the value from the last row of 'data'}

\item{method}{String defining the method to be used by 'optim' to maximise the log-likelihood: default is 'BFGS'}
}
\value{
List of attributes & parameters defining a nonstationary model
}
\description{
Use maximum likelihood estimation to find the parameters of the nonstationary model that best describes a vector of data 'x' using a set of given covariates
}
\details{
Further details to be added for dist & type
}
