% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-parameters.R
\name{get.ns_pars}
\alias{get.ns_pars}
\title{Support function to get location of nonstationary distribution at specified values of the covariates,
given the form and parameters of the nonstationary model}
\usage{
get.ns_pars(fittype, pars, fixed_cov)
}
\arguments{
\item{fittype}{string defining the type of model fit to use - method is currently implemented for 'shift' and 'fixeddisp'}

\item{pars}{vector of named parameters of the nonstationary model}

\item{fixed_cov}{data.frame containing values of the covariate(s) at which the location is to be evaluated}
}
\value{
Return a list containing the location, scale and (if applicable) shape parameters corresponding to the given covariates
}
\description{
Support function to get location of nonstationary distribution at specified values of the covariates,
given the form and parameters of the nonstationary model
}
\keyword{internal}
