% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{getsynchi2}
\alias{getsynchi2}
\title{Support function to calculate the chi^2 function used to estimate sig_mod}
\usage{
getsynchi2(data, sig_mod = 0)
}
\arguments{
\item{data}{Data.frame with one row per climate model, with columns containing best estimate and upper and lower bound of 95pc confidence interval for the quantity of interest}

\item{sig_mod}{Scalar: model representation error. Default value is 0}
}
\value{
Vector of best estimate and 95pcconfidence bounds for weighted model mean, with and without model representation error
}
\description{
Support function to calculate the chi^2 function used to estimate sig_mod
}
\keyword{internal}
