% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{infer_infinite}
\alias{infer_infinite}
\title{Replace infinite probability ratios with finite estimates (for plotting only)}
\usage{
infer_infinite(models_in)
}
\arguments{
\item{models_in}{Data.frame with one row per dataset, with columns "est", "lower" and "upper" giving the best estimate and lower and upper bounds for the quantity of interest. Dataset names should be given as rownames.}
}
\value{
Data.frame with the same models and rows, with infinite values replaced with finite values (where possible).
}
\description{
Replace infinite probability ratios with finite estimates (for plotting only)
}
\details{
Infinite probability ratios are replaced with finite values based on the bounds and estimates from other models, following the approach
proposed in Otto et al. (2024). Where the lower bound and best estimate are both finite, the upper bound is assumed to be a six-sigma interval above the best estimate.
For models where only the lower bound is finite, the highest finite upper bound from the other models is used as the 'best estimate', and the upper bound is inferred in the same way.
Zero lower bounds are not yet handled by this method.
}
\references{
{
Otto, F. E., Barnes, C., Philip, S., Kew, S., van Oldenborgh, G. J., & Vautard, R. (2024). Formally combining different lines of evidence in extreme-event attribution. Advances in Statistical Climatology, Meteorology and Oceanography, 10(2), 159-171.
}
}
