% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing-prob-or-intensity.R
\name{int_change}
\alias{int_change}
\title{Estimate the change in intensity of events of a specified return period under a given model, between the counterfactual and factual worlds}
\usage{
int_change(mdl, rp = NA, cov_f, cov_cf, relative = F)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{rp}{Scalar: return period of interest}

\item{cov_f}{Data.frame of covariate values defining the factual climate}

\item{cov_cf}{Data.frame of covariate values defining the counterfactual climate}

\item{relative}{Boolean: return change as relative (F) or absolute (T)? Default is F}
}
\value{
Scalar representing the estimated change in intensity
}
\description{
Estimate the change in intensity of events of a specified return period under a given model, between the counterfactual and factual worlds
}
