% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate-modelling.R
\name{joint_returnperiod}
\alias{joint_returnperiod}
\title{Get joint return period of two marginal events in a specified climate}
\usage{
joint_returnperiod(joint_model, fixed_cov, ev_x, ev_y)
}
\arguments{
\item{joint_model}{A list containing two nonstationary marginal distributions and a joint copula, as returned by 'fit_copula'.}

\item{fixed_cov}{Data.frame with rows specifying the factual/counterfactual climates for which the joint return period is to be estimated.}

\item{ev_x}{(Optional) scalar: the event value to be evaluated for the first marginal distribution. Default is to use the event value specified during model fitting.}

\item{ev_y}{(Optional) scalar: the event value to be evaluated for the second marginal distribution. Default is to use the event value specified during model fitting.}
}
\value{
Scalar value per row of 'fixed_cov', givving the joint return period according to the specified bivariate model.
}
\description{
Get joint return period of two marginal events in a specified climate
}
