% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_ts}
\alias{load_ts}
\title{Load a time series in Climate Explorer .dat format}
\usage{
load_ts(fnm, ...)
}
\arguments{
\item{fnm}{String defining filename to be loaded}

\item{...}{Additional parameters to be passed to read.csv (for example, used to specify col.names)}
}
\value{
Data.frame containing the .csv data
}
\description{
Load a time series in Climate Explorer .dat format
}
