% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing-prob-or-intensity.R
\name{map_from_u}
\alias{map_from_u}
\title{Support function to map a series of exceedance probabilities u to event magnitudes x under a given model}
\usage{
map_from_u(mdl, u, fixed_cov = NA)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{u}{Vector of exceedance probabilities to be transformed}

\item{fixed_cov}{data.frame containing values of the covariate to be used to transform the probabilities. Default value is NA, in which case nonstationary parameters are returned for all values of the covariate(s) used to fit the model}
}
\value{
Vector of event magnitudes
}
\description{
Support function to map a series of exceedance probabilities u to event magnitudes x under a given model
}
\keyword{internal}
