% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing-prob-or-intensity.R
\name{map_to_u}
\alias{map_to_u}
\title{Support function to map a series of events x to exceedance probabilities under a given model}
\usage{
map_to_u(mdl, x, fixed_cov = NA)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{x}{Vector of event magnitudes to be transformed}

\item{fixed_cov}{data.frame containing values of the covariate to be used to transform the event magnitudes. Default value is NA, in which case nonstationary parameters are returned for all values of the covariate(s) used to fit the model}
}
\value{
Vector of exceedance probabilities
}
\description{
Support function to map a series of events x to exceedance probabilities under a given model
}
\keyword{internal}
