% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-results.R
\name{mdl_ests}
\alias{mdl_ests}
\title{Get best estimates of parameters from fitted model}
\usage{
mdl_ests(mdl, cov_f, cov_cf, ev, rp = NA)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{cov_f}{Data.frame with one row containing named covariates defining the factual climate}

\item{cov_cf}{Data.frame with one or more rows containing named covariates defining the counterfactual climate}

\item{ev}{(Optional) scalar: magnitude of the event of interest. If not provided, event value is picked up from the fitted model}

\item{rp}{(Optional) scalar: fixed return period of interest}
}
\value{
Vector containing estimates of all model parameters and quantities of interest
}
\description{
Get best estimates of parameters from fitted model
}
