% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{ns_loglik}
\alias{ns_loglik}
\title{Support function to calculate the log-likelihood of a vector of data 'x' for a nonstationary distribution defined by the parameters 'pars' (used by 'fit_ns' to estimate the model parameters)}
\usage{
ns_loglik(pars, cov, x, dist, fittype)
}
\arguments{
\item{pars}{Vector of named parameters of the nonstationary model}

\item{cov}{Data.frame containing values of the covariate(s) at which the location is to be evaluated}

\item{x}{Vector of data to evaluate log-likelihood}

\item{dist}{String defining the parametric form to be fitted (currently only implemented for 'gev' and 'norm')}

\item{fittype}{String defining the type of model fit to use - method is currently implemented for 'shift' and 'fixeddisp'}
}
\value{
Scalar log-likelihood
}
\description{
Support function to calculate the log-likelihood of a vector of data 'x' for a nonstationary distribution defined by the parameters 'pars' (used by 'fit_ns' to estimate the model parameters)
}
\keyword{internal}
