% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-parameters.R
\name{ns_pars}
\alias{ns_pars}
\title{Get location, scale and (if relevant) shape of fitted nonstationary distribution at specified values of the covariates}
\usage{
ns_pars(mdl, fixed_cov = NA)
}
\arguments{
\item{mdl}{list of parameters and attributes defining a fitted model, as returned by 'fit_ns'}

\item{fixed_cov}{data.frame containing values of the covariate at which the location is to be evaluated. Default value is NA, in which case nonstationary parameters are returned for all values of the covariate used to fit the model}
}
\value{
Return a list containing the location, scale and (if applicable) shape parameters corresponding to the given covariates
}
\description{
Get location, scale and (if relevant) shape of fitted nonstationary distribution at specified values of the covariates
}
