% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{panel.cor}
\alias{panel.cor}
\title{Support function to show correlations between variables in panels when using pairs()}
\usage{
panel.cor(x, y, digits = 2, ...)
}
\arguments{
\item{x}{Vector of x values}

\item{y}{Vector of y values}

\item{digits}{Integer: number of decimal places to show correlations to. Default is 2.}

\item{...}{Additional graphical paramaters to be passed to text() (eg. to set colour, font size)}
}
\description{
Support function to show correlations between variables in panels when using pairs()
}
\keyword{internal}
