% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_covtrend}
\alias{plot_covtrend}
\title{Plot fitted trend against a single covariate}
\usage{
plot_covtrend(
  mdl,
  xcov,
  trend_cov = NA,
  ci_cov = NA,
  ci_col = "black",
  ev,
  ev_x,
  rp = c(6, 40),
  add_loess = F,
  loess_col = "forestgreen",
  seed = 42,
  nsamp = 500,
  xlim = NA,
  ylim = NA,
  xlab = NA,
  ylab = NA,
  legend_pos = "topleft",
  main = "",
  lwd = 3
)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{xcov}{String indicating the covariate to plot on the x-axis. Must appear in mdl$cov.}

\item{trend_cov}{Data.frame containing values of the covariates to be used to plot the trend. Default value is NA, in which case the trend is estimated at all values of xcov with all other covariates held at their mean value}

\item{ci_cov}{(Optional) Data.frame containing values of the covariates at which confidence intervals for the location parameter should be estimated. Default is NA, in which case no confidence bounds are plotted.}

\item{ci_col}{String: set colour to be used for confidence bounds (if using). Default is 'black'.}

\item{ev}{(Optional) scalar: magnitude of the event of interest. If not provided, event value is picked up from the fitted model}

\item{ev_x}{(Optional) scalar: x-value against which to plot the event of interest. If not provided, event year is picked up from the fitted model}

\item{rp}{(Optional) vector of length two, setting return period for which effective return levels should be plotted. Default is c(6,40)}

\item{add_loess}{Boolean: add a Loess smoother to the plot? Default is F.}

\item{loess_col}{String: set colour to be used for Loess smoother (if using). Default is 'forestgreen'.}

\item{seed}{Scalar: seed to be used to initialise random sample for bootstrapped confidence intervals (if using)}

\item{nsamp}{Scalar: number of bootstrap samples to be used to estimate confidence intervals for location parameter. Set to NA if no confidence intervals are required. Default is 500.}

\item{xlim}{(Optional) vector defining the left and right limits of the x axes}

\item{ylim}{(Optional) vector defining the lower and upper limits of the y axes}

\item{xlab}{(Optional) string: label for x axis. Default is to use 'xcov'}

\item{ylab}{(Optional) string: label for y axis. Default is to leave blank.}

\item{legend_pos}{String indicating location of legend: default is 'topleft'. Change to NA to remove legend.}

\item{main}{String: main title for plot. Default is to leave blank.}

\item{lwd}{Scalar: line weight to be used in plotting. Default is 3}
}
\description{
Plot fitted trend against a single covariate
}
