% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate-plotting.R
\name{plot_joint_contour}
\alias{plot_joint_contour}
\title{Plot contours of equal return periods for a bivariate model.}
\usage{
plot_joint_contour(
  joint_model,
  fixed_cov,
  add = F,
  rp = c(5, 10, 20, 50),
  xlim,
  ylim,
  ...
)
}
\arguments{
\item{joint_model}{A list containing two nonstationary marginal distributions and a joint copula, as returned by 'fit_copula'.}

\item{fixed_cov}{Data.frame with rows specifying the factual/counterfactual climates for which the joint return period is to be estimated.}

\item{add}{Boolean: add to an existing plot? Default is F (create new plot).}

\item{rp}{Vector of scalars: return periods for which contours should be plotted. Default is c(5,10,20,50).}

\item{xlim}{(Optional) vector giving max and min values of x to be plotted. If not provided, will be taken from the range of values in the first marginal model. Ignored if 'add' is T.}

\item{ylim}{(Optional) vector giving max and min values of y to be plotted. If not provided, will be taken from the range of values in the second marginal model. Ignored if 'add' is T.}

\item{...}{Other graphical parameters to be passed to the plotting function.}
}
\description{
Plot contours of equal return periods for a bivariate model.
}
