% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate-plotting.R
\name{plot_joint_event}
\alias{plot_joint_event}
\title{Transform the event of interest into a specified factual/counterfactual climate and plot on existing axes.}
\usage{
plot_joint_event(joint_model, fixed_cov, ...)
}
\arguments{
\item{joint_model}{A list containing two nonstationary marginal distributions and a joint copula, as returned by 'fit_copula'.}

\item{fixed_cov}{Data.frame with rows specifying the factual/counterfactual climates for which the joint return period is to be estimated.}

\item{...}{Other graphical parameters to be passed to the plotting function.}
}
\description{
Transform the event of interest into a specified factual/counterfactual climate and plot on existing axes.
}
