% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_returnlevels}
\alias{plot_returnlevels}
\title{Goodness-of-fit plot comparing return levels in factual and counterfactual climates}
\usage{
plot_returnlevels(
  mdl,
  cov_f,
  cov_cf,
  ev,
  seed = 42,
  nsamp = 500,
  model_desc = T,
  xlim = c(1, 10000),
  ylim = NA,
  pch = 20,
  xlab = "Return period (years)",
  ylab = NA,
  main = "",
  legend_pos = "topright",
  legend_labels = c("Present climate", "Counterfactual climate")
)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{cov_f}{Data.frame with one row containing named covariates defining the factual climate}

\item{cov_cf}{Data.frame with one row containing named covariates defining the counterfactual climate}

\item{ev}{(Optional) scalar: magnitude of the event of interest. If not provided, event value is picked up from the fitted model}

\item{seed}{Scalar: seed to be used to initialise random sample for bootstrapped confidence intervals (if using)}

\item{nsamp}{Scalar: number of bootstrap samples to be used to estimate confidence intervals for location parameter. Set to NA if no confidence intervals are required. Default is 500.}

\item{model_desc}{Boolean: add description of model to plot? Default is T.}

\item{xlim}{Vector defining the lower and upper limits of the x axes (default is c(1,10000))}

\item{ylim}{(Optional) vector defining the lower and upper limits of the y axes}

\item{pch}{Scalar determing the plotting character to be used. Default is 20.}

\item{xlab}{(Optional) string: label for x axis. Default is "Return period (years)".}

\item{ylab}{(Optional) string: label for y axis. Default is to use the covariate name.}

\item{main}{String: main title for plot. Default is to leave blank.}

\item{legend_pos}{String indicating location of legend: default is 'topleft'. Change to NA to remove legend.}

\item{legend_labels}{Vector of labels for legend: default is c("Present climate", "Counterfactual climate").}
}
\description{
Goodness-of-fit plot comparing return levels in factual and counterfactual climates
}
