% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{plot_synthesis}
\alias{plot_synthesis}
\title{Plot synthesis of attribution results from observations and climate models}
\usage{
plot_synthesis(
  synth,
  xlim,
  lwd = 10,
  xlab = "",
  main = "",
  add_space = T,
  log = NA
)
}
\arguments{
\item{synth}{Either a list containing the results of a synthesis carrier out using the 'synthesis' function or a data.frame containing the raw output from a synthesis on the Climate Explorer.}

\item{xlim}{Vector of length 2 (optional) giving range of x coordinates. If not provided, will be estimated from the range of the synthesised values.}

\item{lwd}{Scalar: width of line to be used for each bar. Default is 10.}

\item{xlab}{String with which to label the x axis. Default is an empty string.}

\item{main}{String: main title of plot. Default is an empty string.}

\item{add_space}{Boolean: add a blank row between the observations, models and overall synthesis? Default is T.}

\item{log}{Boolean (optional unless using output from Climate Explorer): should x axis be plotted on a log scale? Default is NA, with value determined by the type of synthesis (PR = T, rel/abs = F).}
}
\description{
Plot synthesis of attribution results from observations and climate models
}
