% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot a fitted trend over time}
\usage{
plot_trend(
  mdl,
  ev,
  ev_year,
  rp = c(6, 40),
  ylab = NA,
  legend_pos = "topleft",
  main = "",
  xlim = NA,
  ylim = NA,
  lwd = 2
)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{ev}{(Optional) scalar: magnitude of the event of interest. If not provided, event value is picked up from the fitted model}

\item{ev_year}{(Optional) scalar: year of the event of interest. If not provided, inferred from the fitted model}

\item{rp}{(Optional) vector of length two, setting return period for which effective return levels should be plotted. Default is c(6,40)}

\item{ylab}{(Optional) string: label for y axis}

\item{legend_pos}{String indicating location of legend: default is 'topleft'. Change to NA to remove legend.}

\item{main}{String: main title for plot. Default is to leave blank.}

\item{xlim}{(Optional) vector defining the limits of the x axis}

\item{ylim}{(Optional) vector defining the lower and upper limits of the y axis}

\item{lwd}{Scalar: line weight to be used in plotting}
}
\description{
Plot a fitted trend over time
}
