% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing-prob-or-intensity.R
\name{prob_ratio}
\alias{prob_ratio}
\title{Estimate the probability ratio of an event under a given model: the factor-change in the probability of exceeding a set threshold when changing from the counterfactual to the factual world}
\usage{
prob_ratio(mdl, ev, cov_f, cov_cf)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{ev}{Scalar: magnitude of event}

\item{cov_f}{Data.frame of covariate values defining the factual climate}

\item{cov_cf}{Data.frame of covariate values defining the counterfactual climate}
}
\value{
Scalar representing how much more or less likely it is that we will exceed the value 'ev' in the factual world vs the counterfactual world
}
\description{
Estimate the probability ratio of an event under a given model: the factor-change in the probability of exceeding a set threshold when changing from the counterfactual to the factual world
}
