% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{refit}
\alias{refit}
\title{Support function to refit the parameters of a model using new data (used in bootstrapping)}
\usage{
refit(mdl, new_data)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{new_data}{Data.frame with named columns containing the variable and any covariates of interest: must contain the same covariates as the data used to fit 'mdl'}
}
\value{
List of attributes & parameters defining a nonstationary model
}
\description{
Support function to refit the parameters of a model using new data (used in bootstrapping)
}
\keyword{internal}
