% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing-prob-or-intensity.R
\name{return_period}
\alias{return_period}
\title{Estimate the return period of an event or events for given covariates}
\usage{
return_period(mdl, x, fixed_cov = NA)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{x}{Vector of data to be transformed}

\item{fixed_cov}{data.frame containing values of the covariate at which the location is to be evaluated. Default value is NA, in which case nonstationary parameters are returned for all values of the covariate(s) used to fit the model}
}
\value{
Vector of return periods
}
\description{
Estimate the return period of an event or events for given covariates
}
