% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{rsquared}
\alias{rsquared}
\title{Calculate R^2 and adjusted R^2 from model output}
\usage{
rsquared(mdl)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}
}
\value{
Vector containing R^2 and adjusted R^2
}
\description{
Calculate R^2 and adjusted R^2 from model output
}
