% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwwa-package.R
\docType{package}
\name{rwwa-package}
\alias{rwwa}
\alias{rwwa-package}
\title{rwwa: Tools for probabilistic attribution following the World Weather Attribution protocol}
\description{
Functions to fit nonstationary models with linear or exponential dependence on covariates of interest, to synthesise the results from fitting similar models to multiple datasets, and to visualise the results.
}
\keyword{internal}
