% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changing-prob-or-intensity.R
\name{stransform}
\alias{stransform}
\title{Transform one nonstationary distribution into another (or into a stationary distribution)}
\usage{
stransform(mdl, fixed_cov = NA)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{fixed_cov}{data.frame containing values of the covariate defining a stationary distribution}
}
\value{
Vector of event magnitudes corresponding to the stationary covariates
}
\description{
Transform one nonstationary distribution into another (or into a stationary distribution)
}
