% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis.R
\name{synthesis}
\alias{synthesis}
\title{Produce synthesis of attribution results from observations and climate models}
\usage{
synthesis(obs_in = NA, models_in, synth_type = "abs")
}
\arguments{
\item{obs_in}{Data.frame with one row per observational dataset, with columns "est", "lower" and "upper" giving the best estimate and lower and upper bounds for the quantity of interest. Dataset names should be given as rownames. Default is NA (no observations), in which case only the model synthesis is carried out.}

\item{models_in}{Data.frame with one row per climate model, with columns "est", "lower" and "upper" giving the best estimate and lower and upper bounds for the quantity of interest. Model names should be given as rownames.}

\item{synth_type}{String defining the type of synthesis to carry out. Options are 'abs' (absolute changes); 'rel' (percentage changes); and 'PR' (probability ratios). Default is 'abs'.}
}
\value{
List containing synth_type; sig_obs (scalar value indicating obs representation error); sig_mod (scalar value indicating model representation error); chi2/dof (initial estimate of ratio of chi^2 to model DOF); and df, a data.frame containing the synthesised results.
}
\description{
Produce synthesis of attribution results from observations and climate models
}
