% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_covtrend}
\alias{plot_covtrend}
\title{Plot fitted trend against a single covariate}
\usage{
plot_covtrend(
  mdl,
  xcov,
  fixed_cov = NA,
  ci_cov = NA,
  ev,
  seed = 42,
  nsamp = 500,
  ylim = NA,
  xlab = NA,
  ylab = NA,
  legend_pos = "topleft",
  main = "",
  lwd = 3
)
}
\arguments{
\item{mdl}{List of attributes & parameters defining a nonstationary model, as returned by 'fit_ns'}

\item{xcov}{String indicating the covariate to plot on the x-axis. Must appear in mdl$cov.}

\item{fixed_cov}{Data.frame containing values of the covariates to be used to plot the trend. Default value is NA, in which case nonstationary parameters are returned for all values of the covariate(s) used to fit the model}

\item{ci_cov}{(Optional) Data.frame containing values of the covariates at which confidence intervals for the location parameter should be estimated. Default is NA, in which case no confidence bounds are plotted.}

\item{ev}{(Optional) scalar: magnitude of the event of interest. If not provided, event value is picked up from the fitted model}

\item{seed}{Scalar: seed to be used to initialise random sample for bootstrapped confidence intervals (if using)}

\item{nsamp}{Scalar: number of bootstrap samples to be used to estimate confidence intervals for location parameter. Set to NA if no confidence intervals are required. Default is 500.}

\item{ylim}{(Optional) vector defining the lower and upper limits of the y axes}

\item{xlab}{(Optional) string: label for x axis. Default is to use 'xcov'}

\item{ylab}{(Optional) string: label for y axis. Default is to leave blank.}

\item{legend_pos}{String indicating location of legend: default is 'topleft'. Change to NA to remove legend.}

\item{main}{String: main title for plot. Default is to leave blank.}

\item{lwd}{Scalar: line weight to be used in plotting. Default is 3}
}
\description{
Plot fitted trend against a single covariate
}
